make the top level directory of the repo a python module
In the top level of the gensim repo you have to
put a a __init__.py containing the following line:
__path__ = './src/gensim'

this points to the actual module. For me it is very useful because I want to
record the commit hashes of the modules I am using for an experiment.

basically it is the answer to the question:
for our experiments we use a framework which records the version of all modules and if they are under version control also the hash of the commit. Usually I have gensim in my home directory and link it to the site-packages. But I link /Users/me/gensim/src/gensim  and the .git directory is in /Users/me/gensim so this framework does not see that the module is under version control. Is there a way to link to /Users/me/gensim and somehow tell python that there is an module in ./src/gensim ?