more efficient sparse matrix generation
When the sparse properties (#documents, #terms, #non-zeroes) are known in advance, a much more efficient code path is taken.
This is the case with MmCorpus, so pass a MmCorpus object to SparseSimilarityIndex whenever possible.
Eligibility for the fast code path is determined by duck-typing, so any corpus supporting self.numDocs, self.numTerms and self.numElements will do (MmCorpus is one such example).