Parameterize LdaModel variational magic numbers
This commit removes the VAR_MAXITER and VAR_THRESH constants from
LdaModel and allows each to be set when a new model is made, or
during a call to update(). They have been named iterations and
gamma_threshold.

Logging messages for the number of passes were changed to use the word
"pass" instead of "iteration", to avoid any confusion. Other logging
messages were also updated to include these vars.

This is a non-automatic fix for Issue #58.