Comment about python version requirements for pattern package (#1011)
* Correcting docs/notebooks/datasets/mycorpus.txt

* Comment about python version requirements

Pattern package which is used in this ipynb is written for Python 2.5+
and has no support for Python 3 yet. Now this is explicitly written in
requirements ipynb section.

* Comment about python version requirement for package pattern

Pattern package which is used in this ipynb is written for Python 2.5+
and has no support for Python 3 yet. Now this is explicitly written in
requirements ipynb section.