matutils.unitvec bug fix proposal
The zeros matrix is coming from the fact that, by nature, a unit vector is going to be less than one, so using vec.dtypes causes use of numpy.int, which rounds all values in the unit vector down to zero. It seems that the only way to stop this from happening (and at the same time maintaining the original bug fix for numpy.float dtypes) is casting before return in order to handle numpy.int types as and when is appropriate, which I have done here. What do you think? (This is my first time using git hub, so please excuse me if I've done something I'm not supposed to do.)