Fix the enumeration over 'alldata'
Without the splitlines(), the enumerate goes over each word rather than each line/sentence. As a result, the split list goes out of range as it encounters values 0,1,2,3, and 4 at which point it throws the list out of bounds error as the number of elements in the list is 4 and the enumerate function gives it values more than 4.