Fixing the computation of the Word2Vec loss.
This commit re-writes the computation of the loss for both CBOW and SG
Word2Vec. The loss that is computed and reported is the running average NCE
loss within the epoch. This means that for each new epoch, the counters are
reset to 0, and the new average is computed. This was not the cas before,
and the loss was incremented during the whole training, which is not
very informative, beside being also incorrect in the implementation (see below)

The computation of the word2vec loss was flawed in many ways:
- race condition on the running_training_loss parameter (updated concurrently in a
  GIL-free portion of the code)
- incorrect dividing factor for the average in the case of SG. The averaging
  factor in the case of SG should not be the effective words, but the effective
  samples (a new variable I introduce), because the loss is incremented as many
  times as there are positive examples that are sampled for an effective word.

Addtionnally, I add the logging of the current value of the loss in the progress
logger, when compute_loss is set to True, and I add a parameter to the
word2vec_standalone script to trigger the reporting of the loss.