Add a link to run tutorial notebooks on Binder
BinderHub is a free (at least for the time being) service that'll automatically provision and run containers that allow notebooks in github repos to be run in a browser with no user installation required. It's a neat project, and I figure this adding this link will lower the barrier to entry for new users.