implement separate functions to load FT embeddings and models (#2376)
Introduced two new pure functions to the gensim.models.fasttext module:

1. load_facebook_vectors: loads embeddings from binaries in FB's fastText .bin format
2. load_facebook_model: loads the full model from binaries in FB's fastText .bin format

The existing FastText.load_fasttext_format method loads full models only. I've placed a deprecation warning around it. The full_model parameter is gone - it was only introduced in 3.7.1, so it's not too late to just rip it out, IMHO.

When releasing 3.7.2, we should include the above in the change log, as it changes the behavior wrt to 3.6.0

Fixes #2372