fixing test_failures that were caused due to a different intitialization of random function
Since we are using a totally different random Generator which is not RandomState, therefore there will be differences in intilizations of weights or any random initialization, than that of last versions. The hardcoded values in tests will fail therfore. I had to change these hardcoded values to the new resluts we get.

Example in test_similarity_mertics , I  added a delta of 5.0e-06 to incorporate small changes.

Note in test_ensemblelda i had to remove 2 tests as these two test were comparing previously saved model with new model , which will be not same as we are using different Random Generator.

I'm not an expert in all the models therefore a review for the changes in test files is required.