Allow overriding the Cython version requirement
Use an environment variable for this since it is often easier to set in a
build wrapper rather than trying to override command-line options in the
right layer of a multi-layer build wrapper and it also requires a lot less
code to do the override.

This will be useful for using alpha versions of Cython or old versions of
Cython provided by the distros or specific versions that fix certain bugs.