Upload the changes to the generated docs to GitHub artifacts
Building the docs often takes too long locally, so this allows
pull request submitters to build on GitHub, download the changes
and incorporate the changes to the generated docs in a commit,
then update their pull request with the generated docs commit.

Also check that the changes to the prebuilt docs are committed,
except for docs that change for every single rebuild.